requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.1
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.1.0
django-countries==7.5.1
django-dramatiq==0.11.6
django-oauth-toolkit==2.3.0
django-phonenumber-field==7.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dramatiq==1.16.0
ecdsa==0.18.0
frozenlist==1.4.0
idna==3.4
jwcrypto==1.5.0
kombu==5.3.5
multidict==6.0.4
oauthlib==3.2.2
phonenumbers==8.13.15
pika==1.3.2
pika-stubs==0.1.3
prometheus_client==0.20.0
prompt-toolkit==3.0.39
psycopg2-binary==2.9.7
pyasn1==0.5.0
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python3-openid==3.2.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.1.0
wcwidth==0.2.6
wrapt==1.15.0
yarl==1.9.2
